# Minor deserialization perf improvements for collections