--- a/src/System.Text.Json/src/System/Text/Json/Serialization/JsonSerializer.Read.Helpers.cs
+++ b/src/System.Text.Json/src/System/Text/Json/Serialization/JsonSerializer.Read.Helpers.cs
@@ -11,11 +11,6 @@ namespace System.Text.Json
             JsonSerializerOptions options,
             ref Utf8JsonReader reader)
         {
-            if (options == null)
-            {
-                options = JsonSerializerOptions.s_defaultOptions;
-            }
-
             ReadStack state = default;
             state.Current.Initialize(returnType, options);