--- a/src/System.Text.Json/src/System/Text/Json/Serialization/ReadStack.cs
+++ b/src/System.Text.Json/src/System/Text/Json/Serialization/ReadStack.cs
@@ -8,7 +8,7 @@ using System.Diagnostics;
 
 namespace System.Text.Json
 {
-    [DebuggerDisplay("Current: ClassType.{Current.JsonClassInfo.ClassType}, {Current.JsonClassInfo.Type.Name}")]
+    [DebuggerDisplay("Path:{JsonPath()} Current: ClassType.{Current.JsonClassInfo.ClassType}, {Current.JsonClassInfo.Type.Name}")]
     internal struct ReadStack
     {
         internal static readonly char[] SpecialCharacters = { '.', ' ', '\'', '/', '"', '[', ']', '(', ')', '\t', '\n', '\r', '\f', '\b', '\\', '\u0085', '\u2028', '\u2029' };
@@ -54,20 +54,17 @@ namespace System.Text.Json
         // Return a JSONPath using simple dot-notation when possible. When special characters are present, bracket-notation is used:
         // $.x.y[0].z
         // $['PropertyName.With.Special.Chars']
-        public string JsonPath
+        public string JsonPath()
         {
-            get
+            StringBuilder sb = new StringBuilder("$");
+
+            for (int i = 0; i < _index; i++)
             {
-                StringBuilder sb = new StringBuilder("$");
-
-                for (int i = 0; i < _index; i++)
-                {
-                    AppendStackFrame(sb, _previous[i]);
-                }
-
-                AppendStackFrame(sb, Current);
-                return sb.ToString();
+                AppendStackFrame(sb, _previous[i]);
             }
+
+            AppendStackFrame(sb, Current);
+            return sb.ToString();
         }
 
         private void AppendStackFrame(StringBuilder sb, in ReadStackFrame frame)