… src/System.Text.Json/src/System/Text/Json/ThrowHelper.Serialization.cs