- Defer calls to ReadStack.JsonPath until an exception needs that information (avoids String and StringBuilder allocs).
- Remove dead code and unused variables


Benchmarks for deserializing single collection `System.Text.Json.Serialization.Tests.ReadJson<ImmutableDictionary<string, string>>`
```
Before
|                   Method |     Mean |     Error |    StdDev |   Median |      Min |      Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------------- |---------:|----------:|----------:|---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
|    DeserializeFromString | 65.60 us | 0.6393 us | 0.5668 us | 65.50 us | 64.64 us | 66.60 us |      6.6489 |      0.5319 |           - |            40.87 KB |
| DeserializeFromUtf8Bytes | 63.73 us | 0.9506 us | 0.8891 us | 63.58 us | 62.40 us | 65.20 us |      5.5444 |      0.5040 |           - |            35.07 KB |
|    DeserializeFromStream | 64.76 us | 1.2135 us | 1.0758 us | 64.87 us | 63.33 us | 66.52 us |      5.6921 |      0.5175 |           - |            35.14 KB

After
|                   Method |     Mean |     Error |    StdDev |   Median |      Min |      Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------------- |---------:|----------:|----------:|---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
|    DeserializeFromString | 64.71 us | 0.8625 us | 0.8067 us | 64.36 us | 63.72 us | 66.20 us |      6.4317 |      0.7718 |           - |            40.75 KB |
| DeserializeFromUtf8Bytes | 62.82 us | 0.6094 us | 0.5701 us | 62.72 us | 62.14 us | 63.98 us |      5.4781 |      0.7470 |           - |            34.94 KB |
|    DeserializeFromStream | 63.31 us | 0.6213 us | 0.5812 us | 63.20 us | 62.42 us | 64.38 us |      5.5710 |      0.7597 |           - |            35.01 KB |
```