@layomia 2019-09-09
I'll create a PR incorporating https://github.com/dotnet/corefx/pull/40889#discussion_r321873298 & applicable pieces from https://github.com/dotnet/corefx/pull/40889#discussion_r321952389 when this PR is merged.